requirements.txt
flask
pdf2docx
camelot-py
gunicorn

